Fixed potential bug with logout function
